refactor code to improve duplicated lines
